Flatten the adapters class hierarchy
