+++ .changeset/thin-geese-worry.md
@@ -1,0 +1,6 @@
+---
+'@keystone-next/keystone-legacy': major
+'@keystone-next/adapter-prisma-legacy': patch
+---
+
+Removed the legacy `BaseKeystoneAdapter`, `BaseListAdapter`, and `BaseFieldAdapter` exports.
