+++ packages/adapter-prisma/src/adapter-prisma.js
@@ -1,11 +1,7 @@
 import fs from 'fs';
 import path from 'path';
+import pWaterfall from 'p-waterfall';
 import { getGenerator, formatSchema } from '@prisma/sdk';
-import {
-  BaseKeystoneAdapter,
-  BaseListAdapter,
-  BaseFieldAdapter,
-} from '@keystone-next/keystone-legacy';
 import { defaultObj, mapKeys, identity, flatten } from '@keystone-next/utils-legacy';
 import {
   runPrototypeMigrations,
@@ -15,9 +11,11 @@ import {
   // eslint-disable-next-line import/no-unresolved
 } from './migrations';
 
-class PrismaAdapter extends BaseKeystoneAdapter {
+class PrismaAdapter {
   constructor(config = {}) {
-    super(...arguments);
+    this.config = { ...config };
+    this.listAdapters = {};
+    this.listAdapterClass = undefined;
     this._prismaClient = config.prismaClient;
 
     this.listAdapterClass = PrismaListAdapter;
@@ -29,6 +27,48 @@ class PrismaAdapter extends BaseKeystoneAdapter {
     this.getDbSchemaName = this.config.getDbSchemaName || (() => 'public');
     this.enableLogging = this.config.enableLogging || false;
     this.url = this.config.url || process.env.DATABASE_URL;
+  }
+
+  newListAdapter(key, adapterConfig) {
+    this.listAdapters[key] = new this.listAdapterClass(key, this, adapterConfig);
+    return this.listAdapters[key];
+  }
+
+  getListAdapterByKey(key) {
+    return this.listAdapters[key];
+  }
+
+  async connect({ rels }) {
+    // Connect to the database
+    await this._connect({ rels }, this.config);
+
+    // Set up all list adapters
+    try {
+      // Validate the minimum database version requirements are met.
+      await this.checkDatabaseVersion();
+
+      const taskResults = await this.postConnect({ rels });
+      const errors = taskResults.filter(({ isRejected }) => isRejected).map(({ reason }) => reason);
+
+      if (errors.length) {
+        if (errors.length === 1) throw errors[0];
+        const error = new Error('Multiple errors in PrismaAdapter.postConnect():');
+        error.errors = errors;
+        throw error;
+      }
+    } catch (error) {
+      // close the database connection if it was opened
+      try {
+        await this.disconnect();
+      } catch (closeError) {
+        // Add the inability to close the database connection as an additional
+        // error
+        error.errors = error.errors || [];
+        error.errors.push(closeError);
+      }
+      // re-throw the error
+      throw error;
+    }
   }
 
   async _prepareSchema(rels) {
@@ -294,10 +334,102 @@ class PrismaAdapter extends BaseKeystoneAdapter {
   }
 }
 
-class PrismaListAdapter extends BaseListAdapter {
-  constructor(key, parentAdapter) {
-    super(...arguments);
+class PrismaListAdapter {
+  constructor(key, parentAdapter, config) {
+    this.key = key;
+    this.parentAdapter = parentAdapter;
+    this.fieldAdapters = [];
+    this.fieldAdaptersByPath = {};
+    this.config = config;
+
+    this.preSaveHooks = [];
+    this.postReadHooks = [
+      item => {
+        // FIXME: This can hopefully be removed once graphql 14.1.0 is released.
+        // https://github.com/graphql/graphql-js/pull/1520
+        if (item && item.id) item.id = item.id.toString();
+        return item;
+      },
+    ];
     this.getListAdapterByKey = parentAdapter.getListAdapterByKey.bind(parentAdapter);
+  }
+
+  newFieldAdapter(fieldAdapterClass, name, path, field, getListByKey, config) {
+    const adapter = new fieldAdapterClass(name, path, field, this, getListByKey, config);
+    adapter.setupHooks({
+      addPreSaveHook: this.addPreSaveHook.bind(this),
+      addPostReadHook: this.addPostReadHook.bind(this),
+    });
+    this.fieldAdapters.push(adapter);
+    this.fieldAdaptersByPath[adapter.path] = adapter;
+    return adapter;
+  }
+
+  addPreSaveHook(hook) {
+    this.preSaveHooks.push(hook);
+  }
+
+  addPostReadHook(hook) {
+    this.postReadHooks.push(hook);
+  }
+
+  onPreSave(item) {
+    // We waterfall so the final item is a composed version of the input passing
+    // through each consecutive hook
+    return pWaterfall(this.preSaveHooks, item);
+  }
+
+  async onPostRead(item) {
+    // We waterfall so the final item is a composed version of the input passing
+    // through each consecutive hook
+    return pWaterfall(this.postReadHooks, await item);
+  }
+
+  async create(data) {
+    return this.onPostRead(this._create(await this.onPreSave(data)));
+  }
+
+  async delete(id) {
+    return this._delete(id);
+  }
+
+  async update(id, data) {
+    return this.onPostRead(this._update(id, await this.onPreSave(data)));
+  }
+
+  async findAll() {
+    return Promise.all((await this._itemsQuery({})).map(item => this.onPostRead(item)));
+  }
+
+  async findById(id) {
+    return this.onPostRead((await this._itemsQuery({ where: { id }, first: 1 }))[0] || null);
+  }
+
+  async find(condition) {
+    return Promise.all(
+      (await this._itemsQuery({ where: condition })).map(item => this.onPostRead(item))
+    );
+  }
+
+  async findOne(condition) {
+    return this.onPostRead((await this._itemsQuery({ where: condition, first: 1 }))[0]);
+  }
+
+  async itemsQuery(args, { meta = false, from = {} } = {}) {
+    const results = await this._itemsQuery(args, { meta, from });
+    return meta ? results : Promise.all(results.map(item => this.onPostRead(item)));
+  }
+
+  itemsQueryMeta(args) {
+    return this.itemsQuery(args, { meta: true });
+  }
+
+  getFieldAdapterByPath(path) {
+    return this.fieldAdaptersByPath[path];
+  }
+
+  getPrimaryKeyAdapter() {
+    return this.fieldAdaptersByPath['id'];
   }
 
   _postConnect({ rels, prisma }) {
@@ -531,10 +663,18 @@ class PrismaListAdapter extends BaseListAdapter {
   }
 }
 
-class PrismaFieldAdapter extends BaseFieldAdapter {
-  constructor() {
-    super(...arguments);
+class PrismaFieldAdapter {
+  constructor(fieldName, path, field, listAdapter, getListByKey, config = {}) {
+    this.fieldName = fieldName;
+    this.path = path;
+    this.field = field;
+    this.listAdapter = listAdapter;
+    this.config = config;
+    this.getListByKey = getListByKey;
+    this.dbPath = path;
   }
+
+  setupHooks() {}
 
   _schemaField({ type, extra = '' }) {
     const { isRequired, isUnique } = this.config;
