+++ packages/keystone/index.js
@@ -1,8 +1,2 @@
 const Keystone = require('./lib/Keystone');
-const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('./lib/adapters');
-module.exports = {
-  Keystone,
-  BaseKeystoneAdapter,
-  BaseListAdapter,
-  BaseFieldAdapter,
-};
+module.exports = { Keystone };
